아이템 상점 UI (2)
- 선택한 아이템 구매 기능 추가
- ItemInfoDb의 Type 자료형을 ItemType에서 Short로 수정

+++ Server/Server/DB/DbTransaction.cs
@@ -51,12 +51,12 @@ namespace Server.DB
 
                             if (info.ItemSlot[i] == null)
                             {
-                                itemInfo.Type = ItemType.None;
+                                itemInfo.Type = (byte)ItemType.None;
                                 itemInfo.Count = 0;
                             }
                             else
                             {
-                                itemInfo.Type = info.ItemSlot[i].Type;
+                                itemInfo.Type = (byte)info.ItemSlot[i].Type;
                                 itemInfo.Count = info.ItemSlot[i].Count;
                             }
                             findPlayer.ItemSlot[i] = itemInfo;
